# Use difflib in hexdiff