The [LCS (wikipedia)](https://en.wikipedia.org/wiki/Longest_common_subsequence_problem) solving algorithm that we are using ([same than on wikipedia](https://en.wikipedia.org/wiki/Longest_common_subsequence_problem#Computing_the_length_of_the_LCS)) is actually also implemented in C in the builtin `difflib` library. See [`difflib.SequenceMatcher`](https://docs.python.org/3.8/library/difflib.html#difflib.SequenceMatcher).

This PR:
- use `difflib` instead our own algorithm
    - For small-sized byte strings, this has little effect. However it REALLY makes a difference for long byte strings.
- provides `autojunk` option that is bundled in `difflib`. It marks some lines as junk to detect the most important changes must faster but will in most cases return a partially wrong diff (for instance, insertions missed and seen as replacements). While it is True by default in `difflib`, we'll therefore use `False` as a default.

### Performance: O(n^2)
```python
import time
def test(n):
    x = time.time()
    a = "A" * n + "findme" + "B" * n
    b = "A" * n + "B" * n
    hexdiff(a, b)
    return "%.3gms" % ((time.time() - x) * 1000)
```

| n   | 10      | 100    | 200    | 500    | 1000  | 10000       |
|-----|---------|--------|--------|--------|-------|-------------|
| OLD | 0.842ms | 42.6ms | 175ms  | 1.1s   | 5.56s | MemoryError |
| NEW | 0.446ms | 4.77ms | 19.6ms | 94.7ms | 355ms | 37s         |

When passing `autojunk=True`, it goes crazy fast (`60ms` for n=10000), but has slightly wrong results. For instance in this example, it shows:
![image](https://user-images.githubusercontent.com/10530980/95030046-f20cd180-06ac-11eb-9ed9-8e01a4bf3936.png)
instead of the proper
![image](https://user-images.githubusercontent.com/10530980/95030066-26808d80-06ad-11eb-9335-62b190b49b1b.png)